fix unit tests with new context getter mock
