+++ src/modules/proxy/proxy-pair/proxy-pair-transactions.service.spec.ts
@@ -10,7 +10,7 @@ import * as Transport from 'winston-transport';
 import { RedisModule } from 'nestjs-redis';
 import { ElrondProxyService } from '../../../services/elrond-communication/elrond-proxy.service';
 import { TransactionsProxyPairService } from './proxy-pair-transactions.service';
-import { ContextServiceMock } from '../../../services/context/mocks/context.service.mocks';
+import { ContextServiceMock } from '../../../services/context/mocks/context.service.mock';
 import { PairService } from 'src/modules/pair/services/pair.service';
 import { PairServiceMock } from 'src/modules/pair/mocks/pair.service.mock';
 import { WrapServiceMock } from 'src/modules/wrapping/wrap.test-mocks';
@@ -22,6 +22,7 @@ import { ProxyService } from '../proxy.service';
 import { ProxyServiceMock } from '../proxy.service.mock';
 import { ProxyPairService } from './proxy-pair.service';
 import { ProxyPairServiceMock } from './proxy.pair.service.mock';
+import { ContextTransactionsService } from 'src/services/context/context.transactions.service';
 
 describe('TransactionProxyPairService', () => {
     let service: TransactionsProxyPairService;
@@ -84,6 +85,7 @@ describe('TransactionProxyPairService', () => {
             providers: [
                 ElrondProxyService,
                 ContextServiceProvider,
+                ContextTransactionsService,
                 ProxyServiceProvider,
                 ProxyPairServiceProvider,
                 PairServiceProvider,
